+++ health-services/individual/src/main/java/org/egov/individual/validators/IdPoolValidatorForUpdate.java
@@ -18,12 +18,14 @@ import org.egov.common.validator.Validator;
 import org.egov.individual.config.IndividualProperties;
 import org.egov.tracer.model.CustomException;
 import org.springframework.core.annotation.Order;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
 
 import static org.egov.common.utils.CommonUtils.isValidPattern;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
 import static org.egov.individual.Constants.*;
+import static org.egov.individual.validators.IdPoolValidatorForCreate.validateDuplicateIDs;
 
 /**
  * Validator class for validating beneficiary IDs during update operations on Individual records.
@@ -37,7 +39,7 @@ import static org.egov.individual.Constants.*;
 @Order(value = 12) // Determines execution order among multiple validators
 public class IdPoolValidatorForUpdate implements Validator<IndividualBulkRequest, Individual> {
 
-    private final BeneficiaryIdGenService beneficiaryIdGenService;
+    private final @Nullable BeneficiaryIdGenService beneficiaryIdGenService;
     private final IndividualProperties individualProperties;
 
     /**
@@ -60,6 +62,8 @@ public class IdPoolValidatorForUpdate implements Validator<IndividualBulkRequest
 
         List<Individual> individuals = request.getIndividuals();
 
+        validateDuplicateIDs(errorDetailsMap, individuals);
+
         // Retrieve existing ID records for the individuals
         Map<String, IdRecord> idRecordMap = IdPoolValidatorForCreate
                 .getIdRecords(beneficiaryIdGenService, individuals, null, request.getRequestInfo());
@@ -78,14 +82,8 @@ public class IdPoolValidatorForUpdate implements Validator<IndividualBulkRequest
 
                 if (identifier != null && StringUtils.isNotBlank(identifier.getIdentifierId())) {
                     String beneficiaryId = identifier.getIdentifierId();
-
-                    if (beneficiaryId.contains("*")) {
-                        // get the last 4 digits
-                        String last4Digits = identifier.getIdentifierId()
-                                .substring(identifier.getIdentifierId().length() - 4);
-                        // regex to check if last 4 digits are numbers
-                        String regex = "[0-9]+";
-                        if (!isValidPattern(last4Digits, regex) || identifier.getIdentifierId().length() != 12) {
+                    if(IdPoolValidatorForCreate.isMaskedId(beneficiaryId)) {
+                        if (!isValidMaskedId(beneficiaryId, individualProperties.getBeneficiaryIdLength())) {
                             updateError(errorDetailsMap, individual, INVALID_BENEFICIARY_ID, "The masked beneficiary id '" + beneficiaryId + "' is invalid.");
                         }
                         continue;
@@ -139,5 +137,18 @@ public class IdPoolValidatorForUpdate implements Validator<IndividualBulkRequest
 
         // Add error to the individual's error map
         populateErrorDetails(individual, error, errorDetailsMap);
+    }
+
+    public static boolean isValidMaskedId(String beneficiaryId, Integer length) {
+        // Check minimum length first
+        if (beneficiaryId == null || beneficiaryId.length() < 4) {
+            return false;
+        }
+        // get the last 4 digits
+        String last4Digits = beneficiaryId
+                .substring(beneficiaryId.length() - 4);
+        // regex to check if last 4 digits are numbers
+        String regex = "[0-9]+";
+        return isValidPattern(last4Digits, regex) && beneficiaryId.length() == length;
     }
 }